+++ MicroStore.Payment.IntegrationEvents/CreatePaymentRequestIntegrationEvent.cs
@@ -7,10 +7,10 @@ namespace MicroStore.Payment.IntegrationEvents
         public string  OrderId { get; set; }
         public string OrderNumber { get; set; }
         public string UserId { get; set; }
-        public decimal SubTotal { get; set; }
-        public decimal ShippingCost { get; set; }
-        public decimal TaxCost { get; set; }
-        public decimal TotalCost { get; set; }
+        public double SubTotal { get; set; }
+        public double ShippingCost { get; set; }
+        public double TaxCost { get; set; }
+        public double TotalCost { get; set; }
         public List<PaymentRequestProductModel> Items { get; set; }
     }
 }
